Made level 6 look a bit more spiffy and cleaned up some code
Added mines around the outer walls of level six to make it seem like it was guarding the home of the boss. Also, I cleaned up some code by adding spacing to my variables to make them seem in sections. Also, I need to make level 6 a bit better in terms of the visual aspects of it. Then, I'd say that I'm done with this game.